Move BaseListener to Shared
