+++ DotNet/Normalization/appsettings.json
@@ -1,6 +1,7 @@
 {
   "ServiceInformation": {
     "Name": "Link Normalization Service",
+    "ServiceName": "Normalization",
     "Version": "1.1.00-dev"
   },
   "EnableSwagger": false,
