+++ DotNet/Shared/Application/Interfaces/IConsumerLogic.cs
@@ -1,7 +1,11 @@
 ﻿using Confluent.Kafka;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace LantanaGroup.Link.DataAcquisition.Application.Interfaces;
-
+namespace LantanaGroup.Link.Shared.Application.Interfaces;
 public interface IConsumerLogic<ConsumeKeyType, ConsumeValueType, ProduceKeyType, ProduceValueType>
 {
     Task executeLogic(ConsumeResult<ConsumeKeyType, ConsumeValueType> consumeResult, CancellationToken cancellationToken = default, params object[] optionalArgList);
