+++ DotNet/Shared/Application/Models/ServiceInformation.cs
@@ -2,6 +2,7 @@
 
 public class ServiceInformation
 {
+    public string ServiceName { get; set; } = string.Empty; 
     public string Name { get; set; } = string.Empty;
     public string Version { get; set; } = string.Empty;
 }
